Windows cleanups.

+++ gitian/mkbundle-windows.sh
@@ -6,7 +6,7 @@
 . ./versions
 
 WRAPPER_DIR=$PWD
-GITIAN_DIR=$PWD/../../gitian-builder.git
+GITIAN_DIR=$PWD/../../gitian-builder
 DESCRIPTOR_DIR=$PWD/descriptors/
 
 if [ ! -f $GITIAN_DIR/bin/gbuild ];
@@ -21,7 +21,12 @@ export PATH=$PATH:$PWD/libexec
 # TODO: Make a super-fresh option that kills the base vms
 if [ ! -f base-precise-i386.qcow2 ];
 then
-  ./bin/make-base-vm --suite precise --arch i386
+  if [ "z$USE_LXC" = "z1" ];
+  then
+    ./bin/make-base-vm --lxc --suite precise --arch i386
+  else
+    ./bin/make-base-vm --suite precise --arch i386
+  fi
 
   if [ $? -ne 0 ];
   then
@@ -31,21 +36,7 @@ then
   stop-target
 fi
 
-if [ ! -d $GITIAN_DIR/inputs ];
-then
-  mkdir -p $GITIAN_DIR/inputs
-fi
-
-#torsocks $DESCRIPTOR_DIR/../fetch-inputs.sh $GITIAN_DIR/inputs
-
 echo "pref(\"torbrowser.version\", \"$TORBROWSER_VERSION\");" > $GITIAN_DIR/inputs/torbrowser.version 
-
-# FIXME: Alpha vs non-alpha xpis??
-cd $GITIAN_DIR/inputs/
-ln -sf $NOSCRIPT_PACKAGE noscript@noscript.net.xpi
-ln -sf $TORBUTTON_PACKAGE torbutton@torproject.org.xpi
-ln -sf $HTTPSE_PACKAGE https-everywhere@eff.org.xpi
-ln -sf $PDFJS_PACKAGE uriloader@pdf.js.xpi
 
 cd $WRAPPER_DIR/..
 zip -rX $GITIAN_DIR/inputs/relativelink-src.zip ./RelativeLink/ 
@@ -55,23 +46,29 @@ zip -rX $GITIAN_DIR/inputs/windows-skeleton.zip ./
 
 cd $GITIAN_DIR
 
-./bin/gbuild --commit tor-launcher=$TORLAUNCHER_TAG,tor-browser=$TORBROWSER_TAG $DESCRIPTOR_DIR/windows/gitian-firefox.yml
-while [ $? -ne 0 ];
-do
-  mv var/build.log ./firefox-fail-win32.log.`date +%Y%m%d%H%M%S`
-  ./bin/gbuild --commit tor-launcher=$TORLAUNCHER_TAG,tor-browser=$TORBROWSER_TAG $DESCRIPTOR_DIR/windows/gitian-firefox.yml
-done
-
-cp -a build/out/tor-browser-gbuilt.zip $GITIAN_DIR/inputs/
-
-./bin/gbuild --commit tor=$TOR_TAG $DESCRIPTOR_DIR/windows/gitian-tor.yml
-while [ $? -ne 0 ];
-do
-  mv var/build.log ./tor-fail-win32.log.`date +%Y%m%d%H%M%S`
+if [ ! -f $GITIAN_DIR/inputs/tor-win32-gbuilt.zip ];
+then
   ./bin/gbuild --commit tor=$TOR_TAG $DESCRIPTOR_DIR/windows/gitian-tor.yml
-done
+  while [ $? -ne 0 ];
+  do
+    mv var/build.log ./tor-fail-win32.log.`date +%Y%m%d%H%M%S`
+    ./bin/gbuild --commit tor=$TOR_TAG $DESCRIPTOR_DIR/windows/gitian-tor.yml
+  done
+  
+  cp -a build/out/tor-win32-gbuilt.zip $GITIAN_DIR/inputs/
+fi
 
-cp -a build/out/tor-gbuilt.zip $GITIAN_DIR/inputs/
+if [ ! -f $GITIAN_DIR/inputs/tor-browser-win32-gbuilt.zip ];
+then
+  ./bin/gbuild --commit tor-launcher=$TORLAUNCHER_TAG,tor-browser=$TORBROWSER_TAG $DESCRIPTOR_DIR/windows/gitian-firefox.yml
+  while [ $? -ne 0 ];
+  do
+    mv var/build.log ./firefox-fail-win32.log.`date +%Y%m%d%H%M%S`
+    ./bin/gbuild --commit tor-launcher=$TORLAUNCHER_TAG,tor-browser=$TORBROWSER_TAG $DESCRIPTOR_DIR/windows/gitian-firefox.yml
+  done
+
+  cp -a build/out/tor-browser-win32-gbuilt.zip $GITIAN_DIR/inputs/
+fi
 
 ./bin/gbuild --commit tbb-windows-installer=$NSIS_TAG $DESCRIPTOR_DIR/windows/gitian-bundle.yml
 while [ $? -ne 0 ];
